+++ Steam/Services/GroupService.cs
@@ -40,21 +40,28 @@ public class GroupService : IGroupServices
     public async Task<IActionResult> Delete(int id, HttpContext context)
     {
         var todelete = await _dbContext.Groups.FindAsync(id);
+        if(todelete == null)
+        {
+            throw new NullReferenceException($"Group was not found under id {id}");
+        }
         if(todelete.Creator == context.User.Identity.Name || context.User.IsInRole("Admin"))
         {
             var userto = await _dbContext.Users.FirstOrDefaultAsync(x => x.UserName == todelete.Creator);
-            await _notificationService.AddNotification(new Notification()
+            if(userto != null)
             {
-                Title = $"Delete From {context.User.Identity.Name}",
-                Description = $"Your group {todelete.Name} was delete by admin",
-                UserTo = userto.Id,
-                UserFrom = context.User.FindFirst(ClaimTypes.NameIdentifier)?.Value,
-                Type = "DeleteGroup"
-            });
-            var notification = await _dbContext.notifications
-                                .OrderByDescending(n => n.Id)
-                                .FirstOrDefaultAsync();
-            await _notificationService.AddNotificationToUser(userto.Id, notification.Id);
+                await _notificationService.AddNotification(new Notification()
+                {
+                    Title = $"Delete From {context.User.Identity.Name}",
+                    Description = $"Your group {todelete.Name} was delete by {context.User.Identity.Name}",
+                    UserTo = userto.Id,
+                    UserFrom = context.User.FindFirst(ClaimTypes.NameIdentifier)?.Value,
+                    Type = "DeleteGroup"
+                });
+                var notification = await _dbContext.notifications
+                                    .OrderByDescending(n => n.Id)
+                                    .FirstOrDefaultAsync();
+                await _notificationService.AddNotificationToUser(userto.Id, notification.Id);
+            }
             _dbContext.Remove(todelete);
             await _dbContext.SaveChangesAsync();
             return new OkResult();
@@ -65,12 +72,20 @@ public class GroupService : IGroupServices
     public async Task<IEnumerable<Group>> GetAll()
     {
         var result = await _dbContext.Groups.ToArrayAsync();
+        if(result == null)
+        {
+            throw new NullReferenceException("Problems in database");
+        }
         return result;
     }
 
     public async Task<Group> GetById(int id)
     {
         var result = await _dbContext.Groups.FindAsync(id);
+        if(result == null)
+        {
+            throw new NullReferenceException($"Group wasn found by id {id}");
+        }
         return result;
     }
 
@@ -80,6 +95,15 @@ public class GroupService : IGroupServices
         return result;
     }
 
+    public async Task<IEnumerable<User>> GetUsersInGroup(int id)
+    {
+        var usersInGroup = await _dbContext.userGroups
+                    .Where(ug => ug.GroupId == id)                  
+                    .Select(ug => ug.User)
+                    .ToArrayAsync();
+        return usersInGroup;
+    }
+
     public async Task<IActionResult> JoinInGroup(int id, string userid)
     {
         if (_dbContext.userGroups.Any(uws => uws.UserId == userid && uws.GroupId == id))
@@ -87,6 +111,10 @@ public class GroupService : IGroupServices
             return new BadRequestResult();
         }
         var toedit = await _dbContext.Groups.FindAsync(id);
+        if (toedit == null)
+        {
+            throw new NullReferenceException("Something went wrong");
+        }
         toedit.MemberCount += 1;
         var toadd = new UserGroups()
         {
@@ -103,6 +131,10 @@ public class GroupService : IGroupServices
         var userWorkShopSubToRemove = await _dbContext.userGroups.FirstOrDefaultAsync(uw => uw.UserId == userid && uw.GroupId == id);
         var toedit = await _dbContext.Groups.FindAsync(id);
         var user = await _dbContext.Users.FindAsync(userid);
+        if(toedit ==  null || user == null || userWorkShopSubToRemove == null)
+        {
+            throw new NullReferenceException("Gorup wasnt found or User");
+        } 
         if(toedit.Creator == user.UserName)
         {
             return new BadRequestResult();
@@ -127,24 +159,33 @@ public class GroupService : IGroupServices
         return result;
     }
 
+    
+
     public async Task<IActionResult> Update(GroupDto dto, int id, HttpContext context)
     {
         var toedit = await _dbContext.Groups.FindAsync(id);
+        if(toedit == null)
+        {
+            throw new NullReferenceException("Group wasnt found in update");
+        }
         if(toedit.Creator == context.User.Identity.Name || context.User.IsInRole("Admin"))
         {
             var userto = await _dbContext.Users.FirstOrDefaultAsync(x => x.UserName == toedit.Creator);
-            await _notificationService.AddNotification(new Notification()
+            if(userto != null)
             {
-                Title = $"Update From {context.User.Identity.Name}",
-                Description = $"Your group {toedit.Name} was update by admin",
-                UserTo = userto.Id,
-                UserFrom = context.User.FindFirst(ClaimTypes.NameIdentifier)?.Value,
-                Type = "UpdateGroup"
-            });
-            var notification = await _dbContext.notifications
-                                .OrderByDescending(n => n.Id)
-                                .FirstOrDefaultAsync();
-            await _notificationService.AddNotificationToUser(userto.Id, notification.Id);
+                await _notificationService.AddNotification(new Notification()
+                {
+                    Title = $"Update From {context.User.Identity.Name}",
+                    Description = $"Your group {toedit.Name} was updated by {context.User.Identity.Name}",
+                    UserTo = userto.Id,
+                    UserFrom = context.User.FindFirst(ClaimTypes.NameIdentifier)?.Value,
+                    Type = "UpdateGroup"
+                });
+                var notification = await _dbContext.notifications
+                                    .OrderByDescending(n => n.Id)
+                                    .FirstOrDefaultAsync();
+                await _notificationService.AddNotificationToUser(userto.Id, notification.Id);
+            }
             toedit.Name = dto.Name;
             toedit.Description = dto.Description;
             toedit.GroupImageUrl = dto.GroupImageUrl;
